app-portage/metagen: Fix 0.7.3 install
Closes: https://bugs.gentoo.org/832069
Package-Manager: Portage-3.0.30, Repoman-3.0.3

+++ app-portage/metagen/metagen-0.7.3.ebuild
@@ -23,9 +23,11 @@ python_install_all() {
 	distutils-r1_python_install_all
 	doman docs/metagen.1
 
-	# Bug 814545
-	mv "${ED}"/usr/share/doc/${P}/* "${ED}"/usr/share/doc/${PF}/ || die
-	rmdir "${ED}"/usr/share/doc/${P}/ || die
+	# Bug 814545 and 832069
+	if [[ ${PF} != ${P} ]]; then  # to be robust across bumps
+		mv "${ED}"/usr/share/doc/${P}/* "${ED}"/usr/share/doc/${PF}/ || die
+		rmdir "${ED}"/usr/share/doc/${P}/ || die
+	fi
 }
 
 python_test() {
